+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `Beaker.dataset.size()`.
 - Added `Beaker.dataset.rename()`.
 - Added `ExperimentSpec.from_file()`.
+- Added `Beaker.cluster.filter_available()` method.
 
 ## [v0.5.3](https://github.com/allenai/beaker-py/releases/tag/v0.5.3) - 2022-04-08